Move remaining types from 'core' to 'data', clean up CMakeLists.txt for test
