+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 # CMake version and project name
-cmake_minimum_required(VERSION 3.5...3.31)
+cmake_minimum_required(VERSION 3.24...4.3.2)
 set(CUTIL_LIBRARY "cutil")
 project("${CUTIL_LIBRARY}" LANGUAGES C)
