+++ lib/data/src/generic/list.c
@@ -1,9 +1,9 @@
 #include <cutil/data/generic/list.h>
 
 #include <cutil/core/string/builder.h>
-#include <cutil/core/string/util/iterator.h>
 #include <cutil/core/util/compare.h>
 #include <cutil/core/util/void.h>
+#include <cutil/data/generic/string/util/iterator.h>
 
 cutil_Bool
 cutil_ListType_equals(const cutil_ListType *lhs, const cutil_ListType *rhs)
