+++ lib/include/cutil/data/generic/string/builder.h
@@ -1,0 +1,24 @@
+/** cutil/data/generic/string/builder.h
+ *
+ * Header for the generic string builder type.
+ */
+
+#ifndef CUTIL_DATA_GENERIC_STRING_BUILDER_H_INCLUDED
+#define CUTIL_DATA_GENERIC_STRING_BUILDER_H_INCLUDED
+
+#include <cutil/data/generic/type.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+ * Generic type descriptor for cutil_StringBuilder.
+ */
+extern const cutil_GenericType *const CUTIL_GENERIC_TYPE_STRING_BUILDER;
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* CUTIL_DATA_GENERIC_STRING_BUILDER_H_INCLUDED */
